Allow many instances to pass the general checker
Now instances that evaluate to being the same in the general case can be
bundled and all passed on from the general typechecker. These most
appropriate single instance will be selected after a language has been
chosen in the concrete generator.

I need to formalize all this. The implementation is going far beyond the
theory. However, the language is not becoming overly complex. The
problem is that edge cases may be missing and the evaluation semantics
are not crystal clear, so it is hard to predict behavior with
confidence.

Next goals:
  [ ] add direct support for C++ STL types to C++ language bindings
  [ ] add numpy support to Python language bindings
  [ ] add raw vector support to R
  [ ] add tests covering all these cases
  [ ] rerun benchmarks
  [ ] rerun case study with R raw vectors