Fix infinite loop bug in Generate.hs
In the code for handling multiple instances, if none of the instances
match the desired type, I need to reduce the type by one step. The
function for this reduction should return Nothing if no reduction occurs
and (Just t) with the reduced type on success. The type should be
evaluated only one step.

I was using the wrong function, one that will return the input type when
no evaluation is possible. This led to an infinite loop. Fixed now, I've
also moved this function to TypeEval, since it is used in both subtyping
and here in Generator.