Add full numpy support to python
Now numpy arrays can serve directly as List implementations.

There is just one case where they are not as first-classy as the python
'list'. This is when a literal list is given, e.g., `foo [1,2,3]`. Here
`[1,2,3]` default to "list". To get a numpy list, we would need to add a
function to convert it to a numpy list. This is not so onerous, since we
would have to do the same in python as well. Also, a hard-coded list
like this is unlikely to be large, so the costs will be minimal.

A downside of this update, is that now numpy is a dependency of morloc
python code, even if the python code does not use numpy. This is
unfortunate, but with a bit of clever hackery, we can probably avoid the
problem. However, for any sort of numeric work, the numpy arrays are
essential, giving order of magnitude improvements of the default list,
so I think the tradeoff is worth it.