Fix python UTF-8 handling and at UTF-8 test
Now Python, R, and C++ all properly handle UTF-8 text. Previously, the
Python language bindings were not including the array size (16 bytes for
the size and relative pointer) in the calculation of the required data
size for a string. This caused too few bytes to be allocated in shared
memory. Then the string overwrote the final block, leading to a "Bad
BLK_MAGIC" warning. The program did not crash, but since it could not
find a free block in the searched volume, it would create a new one.  So
everytime a new string was used, a new and larger memory volume would be
allocated. I should perhaps elevate the BLK_MAGIC warnings to errors
since they indicate a VERY serious bug in the memory management code.