Distributed computing damn it
I made it blazing fast and multi-threaded ... but won't be enough. They
want distributed compute, caching and all that. But chillax babe, I've
got this.

When Morloc was young, and when I was young, I added a means to label a
function. Back then, I intended to add labels, such as effects, caching,
conditions and such. It seems like big then ...

But I digress. The labels may be of use now. I can write a composition
like this:

```
foo = sum . pmap bar
```

This will map in parallel over some job. But I need to specify that
`bar` is the function that needs to be sent off to a new compute node. I
certainly don't want to thow every fucking function off into its own
node. That would be crazy, no self-respecting programmer would do that.
Also, whether or not a function needs to be slammed into solitary
depends on context (e.g., input size). So I can't just say `bar` should
always be run on a compute node. I just want to say *this* `bar` should
go there.

So, labels:

```
foo = sum . pmap big:bar
```

Now many different functions may have the group label `big`. This just
gives the function an arbitrary label, a handle, a typeless meaningless
nothing that is associated with it. But we can give this thingness
meaning.

All that was past, now let's talk about futures. I will add support for
config YAML files (the cool kids are shooting up TOML these days ...
maybe I should to?) that ties runtime info to the labels. I'll name this
config file `local.yaml`:

```
groups:
  big:
    remote:
      threads: 32
      mem: 128
      time: "3-00:00:00"
```

This attaches runtime info to the label "big". When I compile with
default flags, nothing will change. But when I add the flag `--slurm`,
then any function with a label that specifies a `remote` object will
be formed into a remote call. Now for SLURM, this is fairly easy.

Here is what needs doing:
 * add `--slurm` flag to `morloc` compiler
 * parse the local.yaml file and define whatever new records
 * link functions to their labels
 * in the realization step, update scoring since language interop is
   irrelevant across a remote call -- this is lower priority since it
   doesn't affect correctness, only loses some possible optimization
 * add a RemoteCall manifold type, these are similar to ForeignCall
   manifolds, perhaps I should just replace ForeignCall types with a
   more generalized type that may be local (for language interop) or
   remote (for distributed compute)
 * update pool code to make remote calls for RemoteCall manifolds.
   * create a job_config object (a dict in Python) with the job info for
     this remote call (i.e., the `remote` object from the `local.yaml` file)
   * call a special C function `remote_call` (which should be defined in
     the morloc.h file for consistency between languages)
 * `remote_call` takes a call packet and job_config object as a arguments and:
   * make hash for each argument (will need new voidstar hasher function)
   * OR together the hashes and the manifold id and use this to make the
     expected output file where the return packet will be written
   * if this output file already exists in $TMPDIR/cache, then do no
     computation, instead read the file (which should be a raw data
     packet) and return it
   * else, write each argument, serialized to MessagePack, to
     $TMPDIR/cache using the argument hashes as filenames (if the
     filenames already exist, then do nothing, reuse the old values)
   * create a new call packet with every argument having a file
     source and MessagePack format
   * write this call packet directly to a binary file named after the
     required output file with a "call-" prefix (or whatever)
   * use the job description object to start a new SLURM job where the
     script calls nexus.py with three new arguments:
       1. `--packet <packet_path>` - read from packet file
       2. `--packet-lang <lang>` - send to this language pool
       3. `--output raw` - write raw binary packet
       4. `--output-file <file>` - write output to this file
   * wait for the job to finish, then open the output packet
   * if the packet is failing, delete the cached return value
   * return the packet

All this writing to the hard drive is inefficient, but it is needed for
caching anyways, so I guess it is acceptable.

I also need to be able to test this outside of an HPC context. To this
end, I might use the giovtorres/slurm-docker-cluster system to emulate a
cluster through containers.

Anyway, in this commit, I've added a short example (it doesn't work yet).