Add --slurm option to `morloc init`
This adds the flag SLURM_SUPPORT to GCC builds and creates a
`.build-config` file in $MORLOC_HOME that stores whether or not the
current build was made with SLURM support. In the future this file will
store additional compile-time settings.

I've been rethinking the information that should be attached to labels.
These need to be very abstract and they need to be given at the module
level and not inherited. So if a module has the main morloc file
`main.loc`, then it may have a `main.yaml` file that provides the
metadata for all labels in `main.loc`. By requiring the one-to-one
filenames, we enforce the one-to-one relationship. The YAML file might
be like so:

```
group_default:
  benchmark: false
  cache: false
  distribute: null

group:
  big:
    benchmark: true
    cache: true
    distribute:
      threads: 32
      mem: 128
      gpus: 0
      time: "3-00:00:00"
```

This describes only local labels. The `group_default` field would allow
easy toggling of caching, benchmarking, or distributed computing across
the entire workflow. This could be very useful for diagnostics.

The SLURM configuration will generally be the same across all things
running on a cluster. So this should be stored in a file in MORLOC_HOME.
It should define the available queues:

```
slurm:
  queues:
    small:
      threads: 32
      mem: 128
      time: "0-06:00:00"
      gpus: 0
    medium:
      threads: 32
      mem: 128
      time: "0-24:00:00"
      gpus: 0
    large:
      threads: 32
      mem: 128
      time: "7-00:00:00"
      gpus: 0
    gpu:
      threads: 32
      mem: 128
      time: "1-00:00:00"
      gpus: 4
```

This can be automatically generated by `sinfo`. Now the compiler can
choose which queue to use, by finding the smallest ones that meet the
requirements defined for the group.

Next up, I need actually load all these configs and link them to the
function labels. Also I need to update function labels so that a
function can be attached to more than one of them.