Separate remote and foreign in express
Here I need to do something heinous. The tags are associated with terms,
say `bif = someTag:foo . bar`. Here the index of `foo` is associated
with `someTag`. But it is the application index that ultimately is used
as the manifold index. So I need to map the annotation from the head
function to the application. I added the `setManifoldConfig` function
for this purpose.

Then I need to segment based on whether a call is foreign OR remote. A
function is remote if in its `ManifoldConfig` there is a non-`Nothing`
"remote" entry. Previously this field was called "distributed" or
"resources" (the former in the YAML and the latter in the haskell
record). Now I've changed it to remote so that if this field is set it
is implied that the function is intended to be run remotely (I guess
"distributed" had the same meaning, but "resource" did not, since
locally resources might also need to be specified).

Now a `RemoteForm` type, of form:

```
data RemoteForm = ForeignCall | RemoteCall ManifoldConfig
```

Is passed everywhere that PoolCall types (and their homologs) used to be
passed. Whether a function is foreign or remote doesn't matter as far as
segmentation and serialization go. In both cases, data has to be sent
out and read back in from outside the language.

Next up, in the translators I will need to branch to different functions
depending on if a manifold is remote or foreign.