Complete remote wriring down to translators
 * Now foreign and remote calls are separated in the translators. The
   remote calls are just stubs ("REMOTE_CALL").
 * Fixed config parsing
 * Changed build config to YAML

Next I need to implement the remote call functions. For this purpose,
I'll write a custom C function for handling slurm and serialization and
everything. The less that is done in the language-specific pools the
better. I also need an environment where I can test this outside an HPC
architecture.