Reorganize morloc.h with vim folding
Writing all the morloc C code into one header file is a debatable
choice. On the upside, it simplifies the build process. But it leaves us
with huge multi-thousand line files. These large files are not so bad
with a good editor support and organization, though. Here I've
reorganized the file and added in vim fold markers (just set
`foldmethod=marker`).

The immediate motivation for this is that I am about to add few thousand
more lines to the file. This will include porting much of the packet
handling code to here. By doing so, I will have much less repetition
between the language binders (currently they all have to parse their own
packets.

I would also like to move as much of the port handling over here as
well. The more that is maintained in the "morloc.h", the less needs to
be reimplemented across all the languages. This also means I need to
give very careful thought to the "morloc.h" API. For naming conventions,
I should probably error on the side of verbose. Names like "pack" and
"unpack" are much too common and may cause conflicts.

I may also want to move to a one-header and one-source alternative. This
would reduce namespace pollution and improve readability.