Reconsidering endianess
Typically, little-endian is used in compute and big-endian in
networking. The morloc packets are used in sending data over networks,
but they are not expected to be parsed in any special way by network
code. Rather they are read and written by morloc library code.

Big-endian is more readable which is nice for debugging data dumps.
However, it greatly complicates representation in my C code. With
little-endian, I can represent the packet as a struct and directly
represent the integer values -- no special parsing needed.

So I think I should move to little-endian. I am currently working on
moving as much of the packet logic as possible into `morloc.h`. This
will unify logic (rather than the current abomination where my protocol
is defined in every single language binding library AND the nexus).