Considering caching
I am currently working on the C libraries requirements for the big jump
to remote compute support -- well, just slurm for now. One thing I want
to add is caching. This will require hashing. I'm not planning on using
timestamps for caching, rather for every call I will make a hash of the
input arguments (hashing the voidstar data) seeded with the manifold ID
and in the future the hash of the actual function code (or some proxy
for it).

So I need a hashing algorithm. Here are the criteria.

 1. Must be VERY fast - no compromise allowed here, if morloc is to be
    used for scientific compute, it will have to work with the largest
    data sets.

 2. No need for cryptographic guarantees (probably, I need to think this
    through more carefully)

 3. Need a well-supported library

 4. Permissive license and single-header implementation - this allows me
    to package the hashing code with morloc which avoids a dependency

xxhash is great on all points. It is fast and well-maintained. It makes
good use of architecture specific optimizations. It was trivial to set
up and run. I'm happy so far.