Basic sockets are now working with C nexus
Race conditions are tricky in all this socket code. It will probably
take a bit more work to ensure everything is working cleanly and to weed
out rare edge cases.

One very bright spot, though, is that the tricky code will mostly be
relegated to the morloc.h library, not the pools. The new C++ pool is a
meager 200 lines long. And most of this code is fairly soft.

Oh, and the numbers are not being parsed correctly. I suspect this is an
endianess issue.