Basic python example now works
It was quite a pain getting Python multi-processing to work correctly
here. I am still not sure the result is entirely stable. There are edge
cases where everything stalls. Errors in the C code tend to not be
propagated correctly into error packets and instead just freeze the
program. But at least it runs.