Fix morloc argument format auto-detection
If an argument is a readable file, and it has the extensions json or
mpk/msgpack, then it is parsed a JSON or MessagePack, respectively.

If it is readable, but does not have an extension, then it is parsed as
voidstar. If this fails (e.g., it is not smaller than the header length
or lacks the magic momodingding), then the first character is checked to
separate the JSON from MessagePack.

Here there is a slight problem. JSON and MessagePack overlap on the
ascii digits 0-9 (integers 48-56 in MessagePack). But if we assume that
JSON is always newline terminated, then we can separate a FixInt
MessagePack message from a int JSON by length (JSON will never be fewer
than 2 characters). This is a bit sketchy since there is no requirement
that JSON always be newline terminated.

Hand written files (e.g., config files) are usually automatically
newline terminated. Morloc also generates newline terminated JSON. It
should *usually* be fine. This will only be a problem when data is piped
from STDIN or read through process substitution.