Fix R handling, replace poll with pselect
Race conditions are a bitch. I had been using `poll` for streaming data
from the client. But `poll` is non-atomic. When it unblocks signal
input, there is a short window when a message can sneak in before the
polling starts. This led to a weird situation where my code worked when
I added a logging print statment but broke otherwise.