(1/422 fail*) Add full support for gasts
Now call-free code that simply substitutes values into data structures
works for all supported cases.

My approach is to convert the morloc data structure into a JSON string
with C formatting for arguments that need to be substituted. Then the
arguments the user passes are parsed as JSON and the desired fields from
them are extracted and substituted into the data structure.

 TODO:
 * There is a string parsing bug in the `string-encoding` test. It seems
   that I have an extra escape.
 * In the `serial-form-*` tests, the python tests were all hanging.
   Currently all these tests are commented out.