Fixed R asynchronicity
I had somewhere lost the `future` package line where I set the
evaluation to `multicore`. So it was defaulting to `sequential`. This
cause the R pool to hang when it called a foreign pool that made a call
back to R.

There are now two failing tests. The string test fails because I seem to
be double escaping. The `errors` test fails because I am not catching
failure of manifold functions, these failures are causing the pools to
hang.