Replace futures with parallel in R
When a child process is called in R, there is no need for its result to
be returned to the parent. It is passed the client socket file
descriptor and has the responsibility to pass data back to the client.
The R daemon can, like an octopus mommy, just ignore its child after
that. This being the case, I don't really need the extra overhead of
futures. I can just fork a child, cut if off, close the sockets, and
forget about it. This greatly simplifies the R code.

I may need to tweak the parameters of the `parallel` package calls for
optimal performance.

 +------------------+--------------+---------------+---------------+---------------+
 | benchmark        | parallel     | c nexus       | shm           | pre-shm       |
 +------------------+--------------+---------------+---------------+---------------+
 | cZeroBaseline    |    5.6 ± 1.4 | 5.8    ± 1.0  | 52.2   ± 0.7  |   54.1 ± 1.5  |
 | pZeroBaseline    |   47.4 ± 6.1 | 49.2   ± 9.7  | 91.6   ± 3.1  |   96.8 ± 3.5  |
 | pZeroFromForeign |   48.4 ± 6.7 | 51.0   ± 9.9  | 96.2   ± 3.0  |  100.2 ± 1.4  |
 | pZeroToForeign   |   51.4 ± 9.7 | 52.6   ± 10.4 | 99.9   ± 4.1  |  101.3 ± 3.2  |
 | rZeroBaseline    |  383.0 ± 2.3 | 584.0  ± 8.7  | 579.3  ± 3.9  |  580.1 ± 6.2  |
 | rZeroFromForeign |  353.8 ± 0.8 | 583.2  ± 11.7 | 573.0  ± 2.9  |  568.6 ± 3.2  |
 | rZeroToForeign   |  307.8 ± 0.4 | 581.0  ± 3.9  | 571.5  ± 2.8  |  571.4 ± 4.5  |
 +------------------+--------------+---------------+---------------+---------------+
 | cTenBaseline     |   34.5 ± 0.2 | 33.6   ± 0.4  | 82.6   ± 3.0  |   79.7 ± 3.2  |
 | pTenBaseline     |   90.4 ± 7.2 | 88.1   ± 10.2 | 124.0  ± 1.8  |  126.3 ± 4.1  |
 | pTenFromForeign  |  204.0 ± 7.2 | 193.2  ± 9.9  | 238.8  ± 1.6  |  706.5 ± 172  |
 | pTenToForeign    |  168.0 ± 14  | 165.8  ± 10.7 | 217.1  ± 7.8  |  509.0 ± 3.5  |
 | rTenBaseline     | 1032.3 ± 3.2 | 1301.4 ± 7.6  | 1271.3 ± 12.1 | 1266.8 ± 2.9  |
 | rTenFromForeign  |  784.8 ± 8.9 | 1067.2 ± 3.6  | 1033.5 ± 5.6  | 1333.8 ± 122  |
 | rTenToForeign    | 1018.4 ± 13  | 1276.4 ± 3.4  | 1302.1 ± 22.0 | 1538.6 ± 4.8  |
 +------------------+--------------+---------------+---------------+---------------+
 | rMarginalCost1   |  306.9 ± 1.0 | 583.9  ± 11.3 | 580.4  ± 4.5  |  581.7 ± 7.0  |
 | rMarginalCost2   |  311.0 ± 1.1 | 632.5  ± 2.1  | 624.6  ± 2.8  |  621.9 ± 2.9  |
 | rMarginalCost3   |  313.4 ± 1.1 | 724.1  ± 19.3 | 666.6  ± 6.5  |  668.7 ± 18.5 |
 | rMarginalCost4   |  314.6 ± 1.2 | 879.8  ± 14.7 | 741.5  ± 3.3  |  737.9 ± 12.1 |
 | pMarginalCost1   |   47.7 ± 6.2 | 50.6   ± 11.0 | 102.3  ± 0.9  |  108.0 ± 0.9  |
 | pMarginalCost2   |   47.3 ± 3.6 | 49.4   ± 7.6  | 108.6  ± 2.7  |  114.9 ± 1.4  |
 | pMarginalCost3   |   50.9 ± 8.9 | 49.9   ± 6.6  | 111.9  ± 2.8  |  119.2 ± 1.7  |
 | pMarginalCost4   |   50.4 ± 9.1 | 51.0   ± 7.7  | 114.6  ± 1.4  |  122.4 ± 0.7  |
 +------------------+--------------+---------------+---------------+---------------+
 | pMap1000         |   79 ± 1.9   |               |               |  2700         |
 | rMap1000         | 2331 ± 6     |               |               | 23800         |
 +------------------+--------------+---------------+---------------+---------------+

I'm missing a few spots in my map benchmarks. But with the new C
implementation, foreign call overhead has dropped around 10-fold. The
pMap1000 benchmark maps 1000 values from C to Python. It runs in 79ms,
so each call has an overhead of 79 microseconds. I'm getting close the
lower limit UNIX domain sockets set, though I can probably still reduce
this overhead by half. For R, I've dropped overhead to 2ms -- not great,
but still a 10X improvement.

More to the point of this commit message. R costs have dropped by nearly
half. In the `rMarginalCost*` tests, which incrementally add a foreign
call to R in the series:

```
(rid . cid . rnmb)
(rid . cid . rid . cid . rnmb)
...
```
The cost of each new call to R used to be ~100ms, now it is dropped to
~6ms. This is the result of forking the processes rather than
re-starting in new cores. I can probably further improve this is I use
pre-initialized pool of workers. This may come at an upfront cost. This
is especially an issue if the particular workflow only makes serial
calls to R. In the future, it might be worthwhile to do some static
analysis to learn what evaluation strategy is best for a given workflow.