Add remote support
SLURM doesn't quite work yet, but I can know call the `remote_call`
function pools.

 * added `entrypoint` byte to call packets - this is needed to
   distinguish between the parent and worker sides of a remote dispatch
   to the same language. If a particular application should be run on a
   remote machine, then on the parent needs to run one function (say,
   `m42`) and this will call start a job on a remote machine where the
   mirror of this function (say, `m42_remote`) will be run. The same
   pool code is in use on both sides.

 * in pools, dispatch according to the remoteness of the call

 * For RemoteForm, only store the RemoteResources record in the
   RemoteCall constructor.

The `remote_call` function needs the schema for every argument that is
passed to the remote machine. The schemas are required for hashing the
data and reading/writing the packets. But the pool generators do not
have access to the data required to make the schema. Specifically,
Passthrough types are serialized when the enter a pool and are never
used within the pool. They pass through silently and may not even be
representable within the language. Their schema is unknown and without
drastically rewriting the compiler, I can't access this info. A better
solution is to add the schemas to the packet themselves. Self-describing
schemas are worthwile anyway. This may simplify code and make for a
cleaner ecosystem where binary packets can be dumped as JSON (or
whatever) without needing to know their type.

The schema can be written into the metadata region that follows the
packet header. Not much code needs to change for this to work, and no
code needs to change outside of `morloc.h`.