Complete remote_call function
The remote_call function submits a SLURM job. It must take a list of
arguments, hash them, serialize them to cache using their hash names,
make a joint hash from the arguments and the function, check if the
result joint hash is cached, load from cache if possible or otherwise
submit the SLURM job and read the results.

This all compiles now. But the nexus cannot yet support the alternative
path needed in the SLURM worker.

```
sbatch \
  --parsable \
  -o <output_filename> \
  -e <error_filename> \
  <resources> \
  --wrap='<nexus_path> --call-packet <packet_binary_file> --socket-path <socket_path> --output-file <output_file> --output-format raw'",
```

So the nexus needs the new arguments:
  * `--call-packet <packet_binary_file>`: filename of a call packet
  * `--socket-path <socket_path>`: socket file path (e.g., "tmp/pipe-r")
  * `--output-file <output_file>`: filename for output raw data packet
  * `--output-format <output_format>`: output type ("raw" in this case)

What the worker nexus needs to do is fire up the required language
servers and submit the given call packet. Then create the output data
packet. It needs to write the output under a temporary file until it is
fully written, then move to final filename. The parent will be watching
for an output file and will open it as soon as it comes into existence,
so it needs to appear fully formed to avoid being loaded in a partial
state.

I don't currently have a good way to ensure just the required language
servers are activated. So the nexus will just have to fire up all of
them, for now. This isn't a huge problem since these remote calls should
be for pretty time intensive problems, so an extra second of overhead shouldn't
be a problem.