Wrap Python pool manifolds in try-except
This provides a clear traceback

I found the problem (or at least **a** problem) in the remote handling.
In the remote case for calls between languages, there will be two copies
f each manifold, one for the parent and one for the worker. The parent
function runs the special `remote_call` which will ultimately call the
nexus in the remote environment which will create a special "remote"
call packet with the isRemote byte set to true. Wheen this call packet
is evaluated in the remote pool, the job will be dispatched to the
remote version of the manifold (say, "m28_remote"). Or at least, this is
all how it should be. But I have the labels mixed up.

Unfortunately, fixing the mixed up labels isn't so simple. Currently,
the code generator doesn't "know" that a worker manifold is remote; it
only knows that the parent is making a remote call. But the parent
manifold must have the standard name (e.g., "m28") since other functions
do not know or care that it is making a remote call. So I'll need to
flag the remote counterpart when I split them during segmentation and
for it generate the name with the "_remote" suffix. This will require an
update in my data structures and the segmentation code in Generate.hs.