Propagate remote worker versus caller state
Added a HeadManifoldForm type that is initially added to the MonoHead
records that store information about the root manifolds of pools.  These
head manifolds may be ones that are directly called by the user, or they
may be ones that are called only by other pools in foreign or remote
calls. The HeadManifoldForm now stores which of these cases the manifold
meets. For remote calls, this allows the generator to distinguish
between the worker and caller nodes.

I also removed the prior "poolIsRemote" field in the PoolDocs. It is no
longer necessary and wasn't doing its job correctly anyway.