Remote wiring for simple case is now working
All the caching, file creation and packet munging is now working
correctly at least for a simple test case.

I still need to test SLURM. But there is relatively little that would
change in that context. I just need to be sure the shell commands and
associated output parsing are right.

Of course, there is a LOT of work that remains to be done in optimizing
and cleaning all this. I think my memory management is rather bad at the
moment. I do require multiple copies of the full dataset to exist
simultaneously. This could be bad in big data contexts, but is easily
remedied.

The resource and environment management is also very simplistic -- I
just pass the desired walltime, memory, cpu count, and gpu count. But at
least what I have serves as a proof of principle. I can write an
extensive program and cleanly run subgraphs on remote nodes.

So, todo before pushing to master:
 [ ] test this in a slurm ecosystem
 [ ] test with R, Python and C and various combinations with varying numbers of arguments
 [ ] fix whatever is causing the CI run to fail
 [ ] update all containers
 [ ] run all benchmarks