Return correct exit codes from `morloc` commands
Previously the tests would stall when a test failed to compile. This was
because a passing exit code of 0 was returned from `morloc make`
regardless of whether the build succeeded. Now a failing exit code is
returned for any failing `morloc` subcommand. This should fix the
stalled CI builds. Though it does not fix the underlying issue that is
causing the failed builds.