Stop listening for dead children
Previously nexus would wait for a very long time even if the listening
child died immediately (e.g., if there were a syntax error in a python
pool). Now I run waitpid to check the process and reap any that have
zombified.

It might be good to send out the pids to everyone so that all pools can
check that their listeners are alive. Maybe the pid should be sent in
the messages.

I've also dialed down the amount of time that the
send_and_receive_over_socket function waits on a broken/non-existent
socket file. There needs to be some wait time, especially on a laggy
machine, but probably never more than a few seconds. Previously it was
closer to 30 or so.