Update github actions
I need both the numpy module and the numpy C headers. The smart way to
handle this is to ask python where it keeps its headers for numpy and
then pass that path to the C builds. But I'd previously just been
installing the system-wide python with apt-get AND the pip numpy. The
former provides the C heders, but there is no guarantee that the headers
will be up-to-date with the python packages. So not the most stable
thing.