Revert to global types (for now)
Type "aliases" such as `EnglishText = Str` need to be known in any
context where the alias is used (alias here is not quite the right word).
The correct way to distribute the alias would be
through direct descent on the import DAG. But this can get a little
messy. Sometimes a type is needed in a file where the type does not
appear in text. It seems rather ugly to require the user to import types
when they are only used internally by the compiler. A nuclear option is
to make all the type aliases global. Which I foolishly did.

I also added a test, which failed before this commit. It has this
directory structure:

├── lib
│   ├── foo
│   │   ├── foo.py
│   │   └── main.loc
│   └── footypes
│       └── main.loc
└── main.loc

Here `main` imports function `f` from `foo`. `f` has type
`FooType->FooType`. `FooType` is an alias for `Str` and is imported from
`footypes` in `foo`. This should all be correct. In the context where
`f` is defined, `FooType` is in scope. I may actually be misdiagnosing
the problem above. This case should not require global scoping of type
aliases. Following an import DAG, `FooType` should be in scope in main.
Something else is amiss.