Fix JSON parsing bug
Also added a new test covering irregulr JSON formatting and nested
special characters and quotes in strings.

All this bug fixing on the master branch is not ideal. It the tests for
the flu case study and the paper-related debugging are run in containers
that are pulled from master. I need to move these tests to the dev
branch (i.e., have them pull from the latest *dev* branch on github, not
master).