Fix resource exhaustion in C pool
Added a waitpid statement to harvest zombie children. This prevents the
forked processes from accumulating. This accumulation was leading to a
crash after around 19000 calls to C.

My forking solution in C is not ideal. I find that calls to python are
running about 4X faster than calls to C. This is because the python
daemon is using a pool of workers. So it forks the python process
`nprocs` times on initiation and then round robin feeds the jobs to the
child workers as the jobs come.

What I could do is make a proper job queue or thread pool. This would
really speed up my already stupidly overwhelming benchmarks. I'd
probably get overhead down to 10us or so (down from 40us that python
calls currently cost) and pretty close to the domain socket limit.

It would be nice to see the absolute limit of interop speed ...

Once I hit that limit, then I will need to start working on the much
harder problem of reducing memory copying. At the moment, this is the
most concerning performance issue. I need to take data in language A,
refactor it to canonical form A, write it to shared memory, convert it
to canonical form B, and then to its final form in B. If done naively,
this could result in 5 copies of the data. Which, of course, is
perfectly dreadful. If done perfectly, though, I can share memory all
the way across and have only one copy between the two languages.  This
is feasible for tensors, at least.

Anyway ... I guess for demonstration purposes, perhaps I should
implement that C thread pool.