Replace sleep with select in python workers
Previously each of the python workers sleep for a few microseconds,
peek in their pipe to see if any messaged, and then loop back to
sleeping. Now I added a `select` step that will notify them with less
hassle. This doesn't actually translate to any great speed improvement,
but it should reduce chatter.