Slightly improved C++ string voidstar translator
In the paper performance benchmarks with strings of increasing size,
I've found the C-to-Python map is running faster than the Python-to-C
map. This is odd. Timing the foreign_function call in each comparison,
the Python-to-C foreign call is much larger and almost fully explains
the difference. But why?

In this commit, I improved the memory efficiency of input string's
convertion to voidstar (e.g., in put_value functions). This improves
speed. Previously Python put_value functions were running twice as fast
as their C counterparts. Now C is slightly faster. But this change does
not fully solve the discrepancy. There is something else going on.