Fix concrete inference of parameterized type defs
The weaving of general and concrete types into TypeF was failing for
parameterized type definitions (e.g., `type (Foo a) = [(a,Str)]`). Here
the parameterized type `Foo a` cannot be woven element-by-element with
the concrete homolog of `[(a,Str)]`. Instead the `Foo a` needs to be
substituted to its def `[(a,Str)]`, which can be appropriately
woven. Now the inference functions do precisely this.

I've added a test for this that is based on the Fasta case from the
`examples` repo.