Start short dev cycle focused on nexus IO
The main thing I want to add is full and well-tested nexus support for
reading/writing MessagePack and VoidStar.

Currently, it can *read* MessagePack. It can also, in the remote case,
read packets that store MessagePack filenames. But I want to fully
generalize this.

The trickiest part here, by far, is handling VoidStar data. The reason
is that VoidStar contains relative pointers to locations in the shared
memory volumes. These will need to be flattened when we write VoidStar
and re-relativized when we read it.

There are couple other things I could add to nexus IO. One is packet
compression. I could add an option to output compressed JSON data,
packets with compressed data, or compressed MessagePack data. But this
is not a major priority.