Add MessagePack read/write and VoidStar write
Currently the option to set nexus output is not written in usage
statements. I need to fix that.

Also, the nexus can now print flattened VoidStar data. This is done
linearly without memory copying or mutation. It prints to STDOUT, for
now. Might want to generalize to print to any file.

Next I need to add the reading function. This will need to de-flatten
the VoidStar data. It should check the size of the data. Then look for a
single contiguous block that can fit all of it. Then write it to that
block. Then walk through the entire data structure and increment every
relative pointer by the same offset. This should be straightforward
enough.

I also need to test the hell out of it. This is a nice problem becaus I
can test in full circles. Run an identity function that writes to format
X, then read from that output, and see if I get back what I sent
in. Just need to be sure to test thoroughly on deeply nested data.