Add basic usage and arg info nexus help
It's about damn time

Now usage prints as so:

```
$ ./nexus -h
Usage: ./nexus [OPTION]... COMMAND [ARG]...

Nexus Options:
 -h, --help            Print this help message
 -o, --output-file     Print to this file instead of STDOUT
 -f, --output-format   Output format [json|mpk|voidstar]

Exported Commands:
  add
    param 1: Int
    param 2: Int
    return: Int
```

Not perfect. What would be nice would be if the executable took on the
name of the main module.

Next, of course, I need to propagate doc strings. Each function needs a
doc string. The entire module needs a doc string. I also should be able
to print the Morloc version that built the nexus. Maybe also store info
on the environment that produced the executable? Maybe the languages
that are used in each command?

Anyway, this is a start and it already looks a hell-of-a-lot better.